.gitignore: Update to make it more generic
We do not have *.xz, *.gz as part of repositiry, so ignore them. They are
annoying during development period, specially when we want to modify all
changes using `git add -A`.

Also ignore *.swp, and *.rpm because many of such temporary files are
created during development and such files would not be the part of
repository.

Signed-off-by: Pratyush Anand <panand@redhat.com>
Acked-by: Dave Young <dyoung@redhat.com>